moves script things out of modules
more script refactoring


more refactoring


move event detection to a script


logger imports


print becomes logger.info


add logging to sync


linting


fix style


fix circular import


another fix


too many blank lines
