+++ scripts/create_multi_session_mean_df.py
@@ -1,0 +1,14 @@
+import pandas as pd
+from visual_behavior.ophys.io.create_multi_session_mean_df import get_multi_session_mean_df
+
+if __name__ == '__main__':
+    import matplotlib
+
+    matplotlib.use('Agg')
+
+    # cache_dir = r'/allen/programs/braintv/workgroups/nc-ophys/visual_behavior/visual_behavior_pilot_analysis'
+    cache_dir = r'\\allen\programs\braintv\workgroups\nc-ophys\visual_behavior\visual_behavior_pilot_analysis'
+    manifest = pd.read_csv(os.path.join(cache_dir, 'visual_behavior_data_manifest.csv'))
+    experiment_ids = manifest.experiment_id.values
+    get_multi_session_mean_df(experiment_ids, cache_dir,
+                              conditions=['cell_specimen_id', 'change_image_name', 'trial_type'])
