+++ scripts/run_event_detection.py
