+++ tox.ini
@@ -10,7 +10,7 @@ setenv =
     PYTHONPATH = {toxinidir}
 
 commands =
-    pytest -vv
+    pytest -vv -W ignore::DeprecationWarning
 
 deps =
     -rrequirements.txt
