+++ visual_behavior/ophys/io/create_analysis_files.py
@@ -2,24 +2,21 @@ from visual_behavior.ophys.dataset.visual_behavior_ophys_dataset import VisualBe
 from visual_behavior.ophys.response_analysis.response_analysis import ResponseAnalysis
 import visual_behavior.ophys.response_analysis.utilities as ut
 
-import matplotlib
 import logging
 import os
-
-matplotlib.use('Agg')
 
 logger = logging.getLogger(__name__)
 
 
 def create_analysis_files(experiment_id, cache_dir, overwrite_analysis_files=True):
-    print(experiment_id)
+    logger.info(experiment_id)
     logger.info(experiment_id)
     logger.info('saving ', str(experiment_id), 'to', cache_dir)
     dataset = VisualBehaviorOphysDataset(experiment_id, cache_dir)
     analysis = ResponseAnalysis(dataset, overwrite_analysis_files)
 
     logger.info('plotting experiment summary figure')
-    from visual_behavior.ophys.plotting import experiment_summary_figures as esf
+    from visual_behavior.visualization.ophys import experiment_summary_figures as esf
     esf.plot_experiment_summary_figure(analysis, save_dir=cache_dir)
     esf.plot_experiment_summary_figure(analysis, save_dir=dataset.analysis_dir)
     esf.plot_mean_first_flash_response_by_image_block(analysis, save_dir=cache_dir, ax=None)
@@ -32,7 +29,7 @@ def create_analysis_files(experiment_id, cache_dir, overwrite_analysis_files=Tru
 
     logger.info('plotting cell responses')
     save_dir = os.path.join(cache_dir, 'summary_figures')
-    from visual_behavior.ophys.plotting import summary_figures as sf
+    from visual_behavior.visualization.ophys import summary_figures as sf
 
     for cell_specimen_id in dataset.cell_specimen_ids:
         sf.plot_mean_trace_and_events(cell_specimen_id, analysis, ax=None, save=True)
@@ -52,15 +49,3 @@ def create_analysis_files(experiment_id, cache_dir, overwrite_analysis_files=Tru
         sf.plot_mean_response_by_image_block(analysis, cell, save_dir=save_dir)
 
     logger.info('done')
-
-
-if __name__ == '__main__':
-    import sys
-
-    experiment_id = sys.argv[1]
-    cache_dir = r'/allen/programs/braintv/workgroups/nc-ophys/visual_behavior/visual_behavior_pilot_analysis'
-    create_analysis_files(experiment_id, cache_dir, overwrite_analysis_files=True)
-
-    # experiment_id = 747248249
-    # cache_dir = r'\\allen\programs\braintv\workgroups\nc-ophys\visual_behavior\visual_behavior_pilot_analysis'
-    # create_analysis_files(experiment_id, cache_dir, overwrite_analysis_files=True)
