+++ visual_behavior/ophys/response_analysis/response_analysis.py
@@ -4,11 +4,14 @@ Created on Sunday July 15 2018
 @author: marinag
 """
 
+from visual_behavior.ophys.response_analysis import utilities as ut
+
 import os
 import numpy as np
 import pandas as pd
 
-from visual_behavior.ophys.response_analysis import utilities as ut
+import logging
+logger = logging.getLogger(__name__)
 
 
 class ResponseAnalysis(object):
@@ -50,7 +53,7 @@ class ResponseAnalysis(object):
         return path
 
     def generate_trial_response_df(self):
-        print('generating trial response dataframe')
+        logger.info('generating trial response dataframe')
         running_speed = self.dataset.running_speed.running_speed.values
         df_list = []
         for cell_index in self.dataset.cell_indices:
@@ -119,17 +122,17 @@ class ResponseAnalysis(object):
         return trial_response_df
 
     def save_trial_response_df(self, trial_response_df):
-        print('saving trial response dataframe')
+        logger.info('saving trial response dataframe')
         trial_response_df.to_hdf(self.get_trial_response_df_path(), key='df', format='fixed')
 
     def get_trial_response_df(self):
         if self.overwrite_analysis_files:
-            print('overwriting analysis files')
+            logger.info('overwriting analysis files')
             self.trial_response_df = self.generate_trial_response_df()
             self.save_trial_response_df(self.trial_response_df)
         else:
             if os.path.exists(self.get_trial_response_df_path()):
-                print('loading trial response dataframe')
+                logger.info('loading trial response dataframe')
                 self.trial_response_df = pd.read_hdf(self.get_trial_response_df_path(), key='df', format='fixed')
             else:
                 self.trial_response_df = self.generate_trial_response_df()
@@ -170,7 +173,7 @@ class ResponseAnalysis(object):
         return flash_response_df
 
     def save_flash_response_df(self, flash_response_df):
-        print('saving flash response dataframe')
+        logger.info('saving flash response dataframe')
         flash_response_df.to_hdf(self.get_flash_response_df_path(), key='df', format='fixed')
 
     def get_flash_response_df(self):
@@ -179,7 +182,7 @@ class ResponseAnalysis(object):
             self.save_flash_response_df(self.flash_response_df)
         else:
             if os.path.exists(self.get_flash_response_df_path()):
-                print('loading flash response dataframe')
+                logger.info('loading flash response dataframe')
                 self.flash_response_df = pd.read_hdf(self.get_flash_response_df_path(), key='df', format='fixed')
             else:
                 self.flash_response_df = self.generate_flash_response_df()
